flesh out more of the front-end, add some routes for the front-end

+++ app/Http/Controllers/UserFeedsController.php
@@ -5,9 +5,20 @@ namespace App\Http\Controllers;
 use App\Feed;
 use App\UserFeed;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UserFeedsController extends Controller
 {
+    /**
+    * Create a new controller instance.
+    *
+    * @return void
+    */
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -15,7 +26,9 @@ class UserFeedsController extends Controller
      */
     public function index()
     {
-        //
+        $userFeeds = UserFeed::with('feed')->where('user_id', Auth::user()->id)->get();
+
+        return response(json_encode(['status' => 'ok', 'items' => $userFeeds]));
     }
 
     /**
@@ -26,11 +39,14 @@ class UserFeedsController extends Controller
      */
     public function store(Request $request)
     {
-        $feed = Feed::firstOrNew(['url' => $request->input('url')]);
+        $feed = Feed::firstOrCreate(['url' => $request->input('url')]);
 
-        UserFeed::create(array_merge($request->all(), ['feed_id' => $feed->id]));
+        $userFeed = UserFeed::create(array_merge(
+            $request->all(),
+            ['feed_id' => $feed->id, 'user_id' => Auth::user()->id]
+        ));
 
-        return redirect()->back();
+        return response(json_encode(['status' => 'ok', 'items' => [$userFeed]]));
     }
 
     /**
@@ -53,9 +69,13 @@ class UserFeedsController extends Controller
      */
     public function update(Request $request, UserFeed $userFeed)
     {
+        if (Auth::user()->id != $userFeed->user_id) {
+            return response(json_encode(['status' => 'error', 'message' => 'Unauthorized to update this feed']));
+        }
+
         $userFeed->update($request->all());
 
-        return redirect()->back();
+        return response(json_encode(['status' => 'ok', 'items' => [$userFeed]]));
     }
 
     /**
@@ -66,8 +86,12 @@ class UserFeedsController extends Controller
      */
     public function destroy(UserFeed $userFeed)
     {
+        if (Auth::user()->id != $userFeed->user_id) {
+            return response(json_encode(['status' => 'error', 'message' => 'Unauthorized to delete this feed']));
+        }
+
         $userFeed->delete();
 
-        return redirect()->back();
+        return response(json_encode(['status' => 'ok']));
     }
 }
